springboot

+++ src/main/java/com/wms/services/EmplacementServices.java
@@ -7,6 +7,7 @@ import com.wms.repository.EmplacementRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,16 +22,27 @@ public class EmplacementServices {
     }
 
     public Optional<Emplacement> getEmplacemet(final Long id){
+        if(id != null) {
+            return emplacementRepository.findById(id);
+        }
+        return  null;
+    }
 
-        return emplacementRepository.findById(id);
+    public Optional<Emplacement> getEmplacemetByRef(final String ref){
+        if(ref != null) {
+            return emplacementRepository.findByrefemplacement(ref);
+        }
+        return  null;
     }
 
     public void deleteEmplacemet(final Long id) {
-        emplacementRepository.deleteById(id);
+        if(id != null && emplacementRepository.findById(id) != null ) {
+            emplacementRepository.deleteById(id);
+        }
     }
     // -------------   Create
     public Emplacement saveEmplacement(Emplacement emplacement) {
-        Emplacement emp = emplacementRepository.findByrefemplacement(emplacement.getRefemplacement());
+        Optional<Emplacement> emp = emplacementRepository.findByrefemplacement(emplacement.getRefemplacement());
         if(emp == null){
             return  emplacementRepository.save(emplacement);
         }
@@ -38,6 +50,23 @@ public class EmplacementServices {
 
 
     }
+    //----------------- update
+
+    public void updateEmplacement(final Long id, Emplacement emp) {
+        Optional<Emplacement> e = emplacementRepository.findById( id);
+
+        if(e.isPresent()) {
+            String ref = emp.getRefemplacement();
+                if(ref != null) {
+                    emplacementRepository.updateRefofEmplacement(id,ref);
+                }
+            }
+        else{
+                System.out.println( "aucunne modification !!! ");
+        }
+
+    }
+
